Significant work towards a functional crafting system
Started laying out the menu for the crafting system, and got into the weeds with adding and tweaking methods for crafting recipes, slots for holding requirements, slots for holding items, and a script that distributes items for testing purposes. I've also added a couple of cubes that serve as the item distributer and crafting bench for the test. The actual crafting interface still needs to be set up.