Crafting Menu correctly searches for items, traits, and sets recipe
Recent test shows that it was able to populate the first row of the recipe builder correctly by searching the inventory for items with parts with the right trait, assembling those items, and instantiating objects that can display properties from the part and the item. This should allow us to see each part of the same item that satisfies a required trait as "seperate" objects, though they are from the same item.
Next steps are to set up navigation, selection, and actual crafting.