Crafting menu can now store accepted parts and check if the recipe is satesfied
The crafting menu can store which parts are accepted now, and can determine if it meets the requirements of the recipe with the accepted parts. It's pretty buggy. The item and part images aren't changing colors when accepted or occupied yet, and the crafting menu doesn't re-set when we close and re-open it yet.