added cancel behavior and improved occupied behavior
Occupied parts can no longer be confirmed, however there are some issues with canceling or closing the menu that are causing problems with the interface & the inventory system. The two issues to iron out is that closing the menu prevents the parts from being "unaceppted" properly, and that the first slot is not properly selected in the menu when closed and re-opened.