Close, but Check Conditions system hitting a snag
as it stands, I'm using a list of current parts, but the fact that the list isn't a fixed length is causing issues. I may want to try an array here instead, with a fixed length, and check-by-index. at conditions. Otherwise things are coming along nicely. (there is a bug where clicking the 2nd or 3rd slot and slotting something throws errors).